+++ src/cmd/pulls.c
@@ -58,7 +58,7 @@ usage(void)
 	fprintf(stderr, "usage: gcli pulls create [-o owner -r repo] [-f from]\n");
 	fprintf(stderr, "                         [-t to] [-d] [-l label] pull-request-title\n");
 	fprintf(stderr, "       gcli pulls [-o owner -r repo] [-a] [-A author] [-n number]\n");
-	fprintf(stderr, "                  [-L label] [-s]\n");
+	fprintf(stderr, "                  [-L label] [-M milestone] [-s]\n");
 	fprintf(stderr, "       gcli pulls [-o owner -r repo] -i pull-id actions...\n");
 	fprintf(stderr, "OPTIONS:\n");
 	fprintf(stderr, "  -o owner        The repository owner\n");
@@ -66,6 +66,7 @@ usage(void)
 	fprintf(stderr, "  -a              Fetch everything including closed and merged PRs\n");
 	fprintf(stderr, "  -A author       Filter pull requests by the given author\n");
 	fprintf(stderr, "  -L label        Filter pull requests by the given label\n");
+	fprintf(stderr, "  -M milestone    Filter pull requests by the given milestone\n");
 	fprintf(stderr, "  -d              Mark newly created PR as a draft\n");
 	fprintf(stderr, "  -f owner:branch Specify the owner and branch of the fork that is the head of a PR.\n");
 	fprintf(stderr, "  -l label        Add the given label when creating the PR\n");
@@ -552,6 +553,10 @@ subcommand_pulls(int argc, char *argv[])
 		  .has_arg = no_argument,
 		  .flag    = NULL,
 		  .val     = 'L' },
+		{ .name    = "milestone",
+		  .has_arg = no_argument,
+		  .flag    = NULL,
+		  .val     = 'M' },
 		{ .name    = "sorted",
 		  .has_arg = no_argument,
 		  .flag    = NULL,
@@ -576,7 +581,7 @@ subcommand_pulls(int argc, char *argv[])
 	};
 
 	/* Parse commandline options */
-	while ((ch = getopt_long(argc, argv, "+n:o:r:i:asA:L:", options, NULL)) != -1) {
+	while ((ch = getopt_long(argc, argv, "+n:o:r:i:asA:L:M:", options, NULL)) != -1) {
 		switch (ch) {
 		case 'o':
 			owner = optarg;
@@ -611,6 +616,9 @@ subcommand_pulls(int argc, char *argv[])
 		} break;
 		case 'L': {
 			details.label = optarg;
+		} break;
+		case 'M': {
+			details.milestone = optarg;
 		} break;
 		case 's': {
 			flags |= OUTPUT_SORTED;