+++ src/github/pulls.c
@@ -52,11 +52,6 @@
  * the actual storage and instead just record the new size of the
  * allocation. */
 
-struct filter_params {
-	char const *author;
-	char const *label;
-};
-
 static bool
 pull_has_label(gcli_pull const *p, char const *const label)
 {
@@ -68,43 +63,44 @@ pull_has_label(gcli_pull const *p, char const *const label)
 }
 
 static void
-github_pulls_filter_author(gcli_pull **listp, size_t *sizep, struct filter_params *p)
+github_pulls_filter(gcli_pull **listp, size_t *sizep,
+                    gcli_pull_fetch_details const *details)
 {
 	for (size_t i = *sizep; i > 0; --i) {
 		gcli_pull *pulls = *listp;
+		gcli_pull *pull = &pulls[i-1];
 		bool should_remove = false;
 
-		if (p->author && strcmp(p->author, pulls[i-1].author))
+		if (details->author && strcmp(details->author, pull->author))
 			should_remove = true;
 
-		if (p->label && !pull_has_label(&pulls[i-1], p->label))
+		if (details->label && !pull_has_label(pull, details->label))
+			should_remove = true;
+
+		if (details->milestone && pull->milestone &&
+		    strcmp(pull->milestone, details->milestone))
 			should_remove = true;
 
 		if (should_remove) {
-			gcli_pull_free(&pulls[i - 1]);
+			gcli_pull_free(pull);
 
-			memmove(&pulls[i - 1], &pulls[i], sizeof(*pulls) * (*sizep - i));
+			memmove(pull, &pulls[i], sizeof(*pulls) * (*sizep - i));
 			*listp = realloc(pulls, sizeof(*pulls) * (--(*sizep)));
 		}
 	}
 }
 
 static int
-github_fetch_pulls(gcli_ctx *ctx, char *url, char const *const filter_author,
-                   char const *const filter_label, int max,
+github_fetch_pulls(gcli_ctx *ctx, char *url,
+                   gcli_pull_fetch_details const *details, int max,
                    gcli_pull_list *const list)
 {
-	struct filter_params params = {
-		.author = filter_author,
-		.label = filter_label,
-	};
-
 	gcli_fetch_list_ctx fl = {
 		.listp = &list->pulls,
 		.sizep = &list->pulls_size,
 		.parse = (parsefn)(parse_github_pulls),
-		.filter = (filterfn)(github_pulls_filter_author),
-		.userdata = &params,
+		.filter = (filterfn)(github_pulls_filter),
+		.userdata = details,
 		.max = max,
 	};
 
@@ -131,8 +127,7 @@ github_get_pulls(gcli_ctx *ctx, char const *owner, char const *repo,
 	free(e_owner);
 	free(e_repo);
 
-	return github_fetch_pulls(ctx, url, details->author, details->label,
-	                          max, list);
+	return github_fetch_pulls(ctx, url, details, max, list);
 }
 
 int